move unit tests to separate project
(IIRC this wasn't possible before due to the tight coupling)
