+++ TRENZ.Lib.RazorMail.sln
@@ -14,6 +14,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TRENZ.Lib.RazorMail.System.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TRENZ.Lib.RazorMail.MailKit", "TRENZ.Lib.RazorMail.MailKit\TRENZ.Lib.RazorMail.MailKit.csproj", "{8F5F2E31-8A89-424A-B024-401905E2531D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TRENZ.Lib.RazorMail.Tests", "TRENZ.Lib.RazorMail.Tests\TRENZ.Lib.RazorMail.Tests.csproj", "{819186AF-5970-4666-8403-629D44D05A8E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -36,5 +38,9 @@ Global
 		{8F5F2E31-8A89-424A-B024-401905E2531D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F5F2E31-8A89-424A-B024-401905E2531D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8F5F2E31-8A89-424A-B024-401905E2531D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{819186AF-5970-4666-8403-629D44D05A8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{819186AF-5970-4666-8403-629D44D05A8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{819186AF-5970-4666-8403-629D44D05A8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{819186AF-5970-4666-8403-629D44D05A8E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
